+++ src/main/java/com/prowidesoftware/swift/model/field/OptionDPartyField.java
@@ -15,9 +15,6 @@
  */
 package com.prowidesoftware.swift.model.field;
 
-import com.prowidesoftware.deprecation.DeprecationUtils;
-import com.prowidesoftware.deprecation.ProwideDeprecated;
-import com.prowidesoftware.deprecation.TargetYear;
 import java.util.*;
 import org.apache.commons.lang3.StringUtils;
 
@@ -44,13 +41,6 @@ import org.apache.commons.lang3.StringUtils;
  */
 public abstract class OptionDPartyField extends Field implements PartyIdentifier {
     public static final String PARSER_PATTERN = "[[/c][/S]$]S[$S]0-3";
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    public static final String COMPONENTS_PATTERN = "SSSSSS";
 
     /**
      * Types pattern
@@ -152,17 +142,6 @@ public abstract class OptionDPartyField extends Field implements PartyIdentifier
         }
         // default format (as is)
         return getComponent(component);
-    }
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    @Override
-    public final String componentsPattern() {
-        DeprecationUtils.phase3(getClass(), "componentsPattern()", "Use typesPattern() instead.");
-        return typesPattern();
     }
 
     /**