+++ src/main/java/com/prowidesoftware/swift/model/field/OptionRPartyField.java
@@ -15,9 +15,6 @@
  */
 package com.prowidesoftware.swift.model.field;
 
-import com.prowidesoftware.deprecation.DeprecationUtils;
-import com.prowidesoftware.deprecation.ProwideDeprecated;
-import com.prowidesoftware.deprecation.TargetYear;
 import java.util.*;
 
 /**
@@ -40,13 +37,6 @@ import java.util.*;
  */
 public abstract class OptionRPartyField extends Field {
     public static final String PARSER_PATTERN = ":S/S/S";
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    public static final String COMPONENTS_PATTERN = "SSS";
 
     /**
      * Types pattern
@@ -134,17 +124,6 @@ public abstract class OptionRPartyField extends Field {
         }
         // default format (as is)
         return getComponent(component);
-    }
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    @Override
-    public final String componentsPattern() {
-        DeprecationUtils.phase3(getClass(), "componentsPattern()", "Use typesPattern() instead.");
-        return typesPattern();
     }
 
     /**