+++ src/main/java/com/prowidesoftware/swift/model/field/StructuredNarrativeField.java
@@ -15,9 +15,6 @@
  */
 package com.prowidesoftware.swift.model.field;
 
-import com.prowidesoftware.deprecation.DeprecationUtils;
-import com.prowidesoftware.deprecation.ProwideDeprecated;
-import com.prowidesoftware.deprecation.TargetYear;
 import com.prowidesoftware.swift.io.writer.FINWriterVisitor;
 import java.util.*;
 
@@ -35,13 +32,6 @@ import java.util.*;
  */
 public abstract class StructuredNarrativeField extends Field implements NarrativeContainer {
     public static final String PARSER_PATTERN = "S";
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    public static final String COMPONENTS_PATTERN = "S";
 
     /**
      * Types pattern
@@ -101,17 +91,6 @@ public abstract class StructuredNarrativeField extends Field implements Narrativ
 
     @Override
     public abstract String getName();
-
-    /**
-     * @deprecated Use {@link #typesPattern()} instead
-     */
-    @Deprecated
-    @ProwideDeprecated(phase4 = TargetYear.SRU2024)
-    @Override
-    public final String componentsPattern() {
-        DeprecationUtils.phase3(getClass(), "componentsPattern()", "Use typesPattern() instead.");
-        return typesPattern();
-    }
 
     /**
      * Returns the field component types pattern